May be continue
I have paused the project here. This is because I am not sure about the
logic behind the project. There are multiple ways to do, but I am not
sure where routing is being done. I would continue when I would be sure
of that. Maybe continue as sdn router.